Moved test helpers to nupic.support.unittest
