+++ src/nupic/support/unittesthelpers/abstract_temporal_memory_test.py
