+++ tests/integration/nupic/algorithms/extensive_tm_test_base.py
@@ -26,7 +26,7 @@ from abc import ABCMeta
 
 from nupic.data.generators.pattern_machine import PatternMachine
 
-from nupic.test.abstract_temporal_memory_test import AbstractTemporalMemoryTest
+from nupic.support.unittesthelpers.abstract_temporal_memory_test import AbstractTemporalMemoryTest
